requirements.txt
PySide6-Essentials
requests>=2.31.0
